[SERVICES] - [INTEGRATION] - processes : rename of files process instead of processes
