+++ services/integration/src/routes/constitution/process.js
@@ -1,5 +1,5 @@
 const Router = require('express').Router();
-const processes = require('../../controllers/constitution/processes');
+const processes = require('../../controllers/constitution/process');
 
 Router.post('/', processes.create);
 Router.get('/', processes.find);
